+++ packages/db/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @3ngram/db
 
+## 0.7.5
+
+### Patch Changes
+
+- Updated dependencies [43a200c]
+  - @3ngram/schema@0.6.4
+
 ## 0.7.4
 
 ### Patch Changes